Add an event listener for modifiying specific requests
By default the ReCAPTCHA response is stored in a textare with the name
g-recaptcha-response and cannot be changed. To be able to use the normal
symfony2 form behaviour, the data is moved to the corresponding form field
if it is available and the controller was properly configured.
For testing purposes the form and field names are currently hardcoded,
but they'll be configurable.